Brinstar: joint table as GrJoint, the acid state's host-sized home, and the anchor reads that crossed symbols
Fighters fell straight through Brinstar. Four separate things, all in
grzebes.c and its ground vars:

- StageData.joints pointed at a u32[8] whose big-endian bytes are the
  s16 {joint, map, jobj index} stream; on a little-endian host every
  pair came out swapped, one wrong joint got linked and the main floor
  (joint 4 under map 6) stayed in joint-local space. Written as GrJoint.
- grZe_AcidState is laid over grZebes_GroundVars5.xC8. It is 0x24 bytes
  on the console and 0x38 here (three 8-byte pointers), so it ran over
  xE8..xF8: the map-7 GObj kept in xF0 was overwritten by the acid's
  material GObj, the frame proc walked an item's joint tree ("root has
  no child") and the acid columns never got their reference points. The
  state is now a real, aligned member on the PC, the fields the console
  shares with its head alias it through an anonymous union, and the
  rest follows it; xF0 also lost its (u32) store.
- The bubble spawn code reads `(u8*) grZe_8049F140 + phase * 0x24 + off`:
  on the console that walks from the two anchor Vec3 pairs into bubble
  entry (phase - 1)'s x/y, because the three symbols are contiguous in
  .bss. Here they are separate globals in linker order and a bubble
  entry holds host pointers. grZe_pc_word() reads the console's byte at
  the console's offset. grZebes_801DC744's base[2]/base[3] are the same
  class (they are grZe_8049F158[0..1]).
- Peach's Castle parked CmSubject* and a GObj in 32-bit ground vars
  shared with Corneria's ints: reads now ask the camera again, the GObj
  store casts to the slot's real width. Venom's xC4 store likewise.

MELEE_MPLINK=1 prints each collision joint's binding; MELEE_STAGE_DIAG
adds Brinstar's column heights, bubble entries and anchor words.

Brinstar lockstep (Captain vs Captain): first divergence 151 -> 395;
the bubble array matches the console word for word through frame 152.

+++ src/melee/gr/grvenom.c
@@ -479,7 +479,7 @@ void grVenom_80203B18(void)
 #if BUILD_TARGET_PC
             gp1->u.venom.xC4 = (uintptr_t) gobj;
 #else
-            gp1->u.venom.xC4 = (u32) gobj;
+            gp1->u.venom.xC4 = (uintptr_t) gobj;
 #endif
             grVenom_80203EAC(9);
             gobj1 = grVenom_80203EAC(7);
